refactor : LottoReward 패키지 변경
