+++ src/main/java/lotto/domain/LottoReward.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public enum LottoReward {
